feat: new data api
The data api is now renamed vaa-data to separate it from the other stuff in lib. The reason is that it should be a standalone module that may in the future be available by itself.

NB. We use ugly "export * from X"-type mass imports, which will be cleaned up later.