feat: use vaa-matching to calculate results
We have a wrapper function for the vaa-matching algorithm in
/utils/matching. When the user lands on /results this is called with the
data stored in localStorage and supplied by server load.

fix: filter answers that don't have matching questions. This is
especially a problem if the Strapi mock data is regenerated. Then we'll
have stored answers with old questionIds.

refactor: move matching calculations to stores