refactor: rename getData calls and stores for better logic
We had mixed nomenclature with, e.g., getAllCandidates. Now, all 'All'
terms are removed unless they have a specific meaning. Currently, the
only place it's used is in the getAllParties API call, which does just
that. We'll only use that internally, however, and call the
getNominatingParties in the frontend instead.