feat: define Tailwind and Daisy classes and colors
Define a limited set of Tailwind utility classes for, e.g., spacing and
font sizes, as well as the colours DaisyUI uses.

The reason for limiting Tailwind classes is that this way adherence to
the design system is easier, because Tailwind gives nice autocomplete
for the available options. For most classes names, such as `md` and `lg`
are used. For spacing numeric values are also available, such as `w-40`,
with the named ones, such as `gap-md`, reserved the most commonly used
values.

Note that you can still use arbitrary values with any Tailwind utility
class with the bracket notation, e.g. `w-[21.35px]`. See:
https://tailwindcss.com/docs/adding-custom-styles#using-arbitrary-values

For DaisyUI, all the basic colours are now defined for both the default
and the dark theme, which is used in dark mode. Note that there's also a
collection of further CSS vars used by DaisyUI contained in
`themeCSSVars`. These are the same for both dark and light themes, hence
the `const`. For more informations, see:
- https://daisyui.com/docs/colors/
- https://daisyui.com/docs/themes/#-5