refactor: apply updated styles to (voter) layout and fix footer
Apply updated Tailwind styles and fix the `footer` so that it shows on
those pages where it was supposed to be shown. Also remove the questions
tip from here and place it on the `questions/` page where it belongs.

Note that the whole `footer` will be removed in the next styling PR so
no further fixes are made.