feat: HeroEmoji component for emoji illustrations
The component styles its content (an emoji) in such a way that it can be
used as an easy illustration on typical pages. The `div` produced is
`aria-hidden` by default but this can be overriden by supplying a
`"false"` value value for that attribute.

The emoji content on each page may be hard-coded or, bettes yet,
supplied using the standard way for text translations.

Fixes #235