feat: Navigation component
The `<Navigation>` component is used to create the navigation menu
contents in page templates. It takes as its content `<NavGroup>`
components which in turn contain `<NavItem>` components. These render as
either `<a>` or `<button>` elements depending on whether the `href`
attribute is supplied.

The component will mainly be used in the page templates.