feat: Icon component
The `Icon` component can be used to display icons in the app with the
icon to use specified by the `name` property. The icon's color and
size as well as other properties of the wrapping `<svg>` may be
specified.

The actual `svg` paths are defined in `ts` files contained in the `svg`
folder's subfolders. Those in the `material` folder use designs from
Material icons whilst  those in `custom` are custom made for the
application. For copyright information, see the description in
`icons.ts`.