feat: shared Color type and safelist fill utility classes
Define a type for DaisyUI color names to be used in components. To be
able to use names for the Tailwind utility classes, such as
``fill-${colorName}``, the fill classes are also safelisted in
`tailwind.config.cjs`. See:
https://tailwindcss.com/docs/content-configuration#safelisting-classes

The colors  are used with the `Icon` component and later in other
components as well. Further classes, e.g. `bg-` or `text-`, may be
safelisted later.