feat: AppLogo template part
The `AppLogo` template part can be used to place the application's
logo. It currently handles switching between light and dark modes and
offers three predefined sizes in which the logo can be used. There's
also an `inverse` property, which will reverse the colors and can be
used when the logo is set on a dark background.

In the future, the component should also allow the use of an image file
defined in Strapi.