feat: Button component
A component for buttons that mostly contain text and an icon. Use the
`variant` prop to specify the button type. When using an `icon`, use
`iconPos` to set the position of the icon relative to the text.

The component can be used to implement most of the buttons in the app
designs, notably the big main action buttons at the bottom of the
screens (`variant="main"`), the secondary actions in the header
(`variant="icon"`) and the sorting and order buttons in results lists
(`variant="normal"` iconPos="left"`).