feat: OpenVAALogo component
The `OpenVAALogo` component displays the OpenVAA logo. You can define
the `color` and `size` of the logo using predefined values. The logo
is rendered as a `<svg>` element.

The component is used on the front page footer. It will be necessary on
other pages in the future as well, such as those containing information
about the app.