feat: prevent non-emoji glyphs in HeroEmoji
Remove all non-emoji characters from `HeroEmoji` content. To do this,
slots can no longer be used, so the content is now supplied in the
`emoji` property. If the property is empty after removal, the component
will not be rendered at all.

Also, increase the `line-height` of `HeroEmoji` to prevent emojis being
clipped at the top.