feat: components for displaying questions
This update creates several components for displaying different parts of
a question, such as `QuestionActions`, which displays the additional
actions available, such as 'Next', 'Previous' and 'Delete answer'.

The purpose for providing these partial components is to show the
questions consistently for both voters and candidates when answering
because any differences in how they're displayed may effect answers.

Furthermore, the `LikertResponseButtons` now supports use both for
answering (for both the voter and a candidate) and for displaying
answers when a voter is looking at a candidate's profile.