fix: fetch single Candidate or Party and clean up getData
The `getData` calls for getting individual Candidates and Parties now
target the correct entity ids and not the Nomination id.

The same logic is also extended to `getElection`.

Additionally, unnecessary test types and functions are removed, as well
as the `addId` utility, which caused issues because it would change the
data returned from Strapi to a single object instead of an Array if
used.