feat: utility for adjusting color contrast
The `color` utilities enable automatically adjusting arbitrary colors,
e.g. those defined for parties or categories, to have high enough
contrast on a background color, such as the 4.5 ratio dictated by WCAG
AA level.

The current implementation is a bit unelegant and uses an iterative
approach to find the proper color value, because coming up with an
analytical solution was too hard for me :D