feat: add party colors defined in Strapi to party properties
Read the user-defined party colors from Strapi and ensure their contrast
using a color-contrast utility. The update anticipates a
yet-to-be-implemented update to the content model, which includes
`color` and `colorDark` properties for both `Parties` and `Question
Categories`. For the latter, however, the definition of the `category`
property of  `QuestionProps` will require a breaking change, so it's
postponed for a later update.