fix: update localisation
Fix issues with localisation, which were mainly caused by fetching the
default locale content without adding a `lang` parameter in the url.

The new localisation paradigm is simpler but now always uses a `lang`
parameter. This currently happens even if there is only one locale
available. This should be corrected in the future.