feat: convert getRoute to a store
Convert the `getRoute` utility to a `derived` store dependent on the
`page` store. This is needed because any links on the page created using
`getRoute` would not update if the url or `lang` route param changes.

After the update, calls in the form `getRoute(foo)` have to be converted
to `$getRoute(foo)`.