fix: localisation of server-loaded data
Fixes localised text updates when the localised data is part of
server-loaded data, including `App Labels`.

Server `load` functions now depend on either the `lang` route param or
the outermost layout's `data.i18n.currentLocale`, which is awaited for
using `parent()`.

When data is loaded on the server and updated due to locale changes,
components using it are not automatically rerendered by the client. This
update makes all such data reactive in the frontend.

The subobjects of `data` used are now made explicit, but a more concise
way would be to remove `export let data` altogether and always call
`$page.data.foo` to get the requisite data.