feat: add question category colors and question ordering to object model
Update the frontend object model to enable `QuestionCategoryProps` in
the `QuestionProps` objects instead of just a string for the category's
name. The new object contains category colors as well as the `order`
property.

Add the `order` property for questions as well and use it along with
that of the category to sort the questions when retrieved from the
database.