fix: resultsAvailable store
Fixes a bug in the `resultsAvailable` store that would make it `false`
before questions had been loaded even though the voter's answers were
available.

In the future, it would be best to load questions earlier and prune the
saved answers so that answers to questions that are not present are
removed.