refactor: implement updated AnswerDict format
The new `vaa-matching` `HasMatchableAnswers` interface expects an
`AnswerDict` type for the answers. We now store all of the answers
already in that format when fetching data by the `getData` utilities.

This affects a number of other types and interfaces and reflects on some
stores and components as well as the `answers` utilities, which are all
updated.