feat: filter stores
The `candidateFilters` store contains the active filter objects. It is
provided as a global store becase we want the filters to persist after
the user has opened a candidate for the detail view.

Also, in the future the active filters are likely needed on other pages,
such as results visualisations.