feat: filtering stores
Add `filteredCandidateRankings` store that can be used to access the
results of filtered `candidateRankings`. This is derived from the
`candidateFilters` store, which in turn contains the active filter
objects.

Global stores are used for these instead of objects local to the results
page, because:

1. We want the filters to persist after the user has opened a candidate
for the detail view
2. In the future accessing the filtered results is most likely needed
when building visualisation, for example.