feat: EntityFilters components
The `<EntityFilters>` component displays filters from the `vaa-filters`
module for filtering entities, e.g. candidates, parties or their
rankings.

The `<TextEntityFilter>` component renders the UI for text-based
filters, such as name.

The `<EnumeratedEntityFilter>` renders an enumerated filter
that displays a list of values to include in the results. These can
be, for example, parties or answers to enumerated questions, like
gender or language. The filter works for both single and multiple
selection questions.

The UI components do not yet implement all of the filters provided by
the `vaa-filters` module, i.e. `NumericFilter`.