feat: get AppSettings from Strapi
Retrieve some settings from Strapi. These are merged with the local
settings defined in `$lib/config/settings.json`, and the effective
settings can be accessed via the `settings` store.

NB. The `stores.ts` file has now a wide variety of stores and in the
future it should be split into multiple files in a `stores` folder.