feat: add a font setting
With the `font` setting, the app's main font can be easily changed. In
order to enable dynamic loading of remote fonts, the font import has to
be moved from `app.css` to a `<link>` element.

For ease of use, the fallback fonts will can only be selected from two
options: `sans` or `serif`. If more customisability is needed, we could
allow instead of providing the raw `font-family` string, e.g. `Inter
Helvetice sans-serif`, and actually it can now be done as well but not
recommended.