refactor: MaybeRanked type for entity components
Components used to display and list entities now accept a single
property `content(s)` for input, which can be used to supply either
naked `EntityProps` or `RankingProps`. This simplifies the use of the
components, because earlier they had two different properties for each
type of input.

The `$lib/utils/entities` module provides helper functions to handle
checking whether a content item is wrapped in `RankingProps` etc.