feat: utility stores for page.data and defaults for candidateRankings
Add derived stores for `page.data.candidates`, `parties`,
`infoQuestions` and `opinionQuestions` so that pages and stores relying
these can be built more easily.

Also, modify `candidateRankings` so that it returns unranked, wrapped
candidates when results are not available for easier use on the
`/results` listing page. (In a separate commit, the redundant
`/candidates` and `/parties` routes are removed and replaced by
`/results`).