feat: use /results route for unranked candidates and parties
Use the same `/results` page for browsing candidates and parties before
rankings are available. This removes a lot of repeated code and logic.

NB. After the update, routes for individual entities, e.g.
`/results/candidate/123` demand all candidates and parties to be loaded,
while the old paradigm allowed for fetching just a single entity. This
is deemed not to be a huge issue, because the most likely use case will
involve the user navigating to the results afterwards, but it can be
corrected later if needed.