deps: isomorphic-dompurify
`Isomorphic-dompurify` is a version of `dompurify` that can be run both
on the server and the client. It is needed to sanitize output that is
rendered using Svelte's `@html` command.

Currently, this is needed for rich text Question info or video
transcripts but there will be more contexts where this is necessary in
the future.