feat: video slot for BasicPage template
Adds a `video` slot to the `<BasicPage>` template which can be used to
display video content at the top of the page.

Also adds a `titleClass` property that can be used to add custom classes
to the `<div>` containing the `title` slot.

Also fixes bug in the `headerClass` property of the `<Page>` template.