feat: custom colors for text Avatar and dynamic text sizing
Adds properties for defining a custom bg color for the `<Avatar>`
component when it's used to show initials, as well as one for defining
the text color.

Also changes the text size dynamically based on the length of the
`initials` text so that it fits better on the background.

NB. For an advanced solution, this might be changed to an `<svg>`
element in the future, which can automatically adjust text size to fit
the allotted space.

The first use case for this is for displaying parties.