refactor: EntityCard
Moves the `action` definition to be handled by the `<EntityCard>`
component and adds a new `subcards` property that can be used to show
nested cards within a card, such as candidates for a party.

Also, moves styling from the `<Card>` component directly to
`<EntityCard>`, which deprecates `<Card>` as it has no other uses. At
the same time the layout is converted into one based on CSS grid.

e   (use "git push" to publish your local commits)