feat: setting for showing a party's candidates in EntityDetails
Adds a `candidates` option for the `entityDetails.party.contents`
setting which if sets lists the parties candidates in the
`<EntityDetails>` card.

Also adds a `showMissingAnswers` setting which controls whether a marker
is shown for missing answers to info question or they are left
completely unrendered. This does not affect opinion questions' display.