feat/deps: enable analytics and data consent popup
Adds analytics and possible data consent popup to the outermost layout
on the `/(voter)` route.

Also adds a `analytics/track` module which can be used to track extended
events. Note that it only acts as an API and the actual function that
sends the data must be supplied to the `sendTrackingEvent` store
contained in the module.

The tracking methods relies on listening to page visibility changes to
which effect the `svelte-visibility-change` component is added to dev
deps.