fix: prevent server crashing when questionId is invalid
(Crawler) requests for an invalid url `/questions/_first_` instead of
`__first__` caused the `error` function to be called in SSR, which
crashes the server.

The update only redirects the user if the id is invalid. This is only
done in CSR because the navigation functions cannot be used in SSR.