feat: add start search param to set a question as first
Enable setting a question as the first on the list by setting a `start`
search param to any value. If present on the
`/(voter)/questions/[questionId]`, the `+page.ts` loaded will set the
`setQuestionAsFirst` `PageData` property to `true`, and the
`+page.svelte` will save the current id (the value of `questionId` route
param, in turn) in a session-persistent store which is used to order the
questions before assigning indeces.

The `firstQuestionId` store is saved in `sessionStorage` so that the
order is not lost on page reload.