feat: don't generate mock App Labels
Disables mock data generation for `App Labels`, because they cannot
currently be easily kept up-to-date with the local translations. In the
future it would be best to define the content type dynamically to always
all appropriate keys from the translations, but this maybe limited to
only those that are mostly frequently edited if exposing all of them is
unwarranted.