chore: test that dynamic translations match
Add a test that checks that the dynamic translation files used by the
frontend and backend match.

This is a temporary fix to ensure that the backend and frontend
translation files match, and will be deprecated when the translations
are moved to a shared module. This test could easily be run with
`vitest`, but because there are currently no other non-e2e global
tests, `playwright` is used.