feat: only show `party` field if there are parties
Show the `party` field on the Candidate App basic info page only if the
are `Party` objects in the database. To check for this, we need to query
for all parties in the `CandidateContext`.

In the future, it might be better to add a custom route to Strapi that
would collect some statistical properties such as this without needing
to explictly fetch the objects of interest themselves. These could be
used elsewhere in the app as well.