feat: load default question types on initialize
Loads some default `Question Type`s to Strapi when initializing. Note
that the default data is now contained within Strapi but it would be
better if it resided in a shared module. See issue #148.

The `loadDefaultData` function only handles question types now, but
other data could also be initialized by default. Another possibility is
to merge this with `loadData` such that whatever data is contained in
the `defaultData` folder is loaded.

Also refactors the default setting and app customisation loading a bit.