chore: update typing and fix linter errors in Strapi
Adds propert types for the access control functions used on the API
routes. The types provided by Strapi have some limitations for which
reason an unseemly `as unknown as Generic` assertion is added to all of
the ACL config in the routes.

Also, fixes some linter errors by mostly adding eslint ignores.