refactor: legacy types
Mark all global types that will be replaced in this refactoring with the
`Legacy` prefix or, in the case of the Candidate App specific types,
mark the type file with the `legacy-` prefix.

These are converted into the nomenclature piecewise so we need to
maintain backwards compatibility for a while for the Candidate App.