feat: remove `strapi-plugin-import-export-entries`
The custom plugin is not compatible with Node 20 or Strapi 5. The
functionalities it provides should be rewritten as a standalone plugin
in the same way as `candidate-admin`.

E2e tests need to also be modified, because they use the import tool to
create the user tested.

This commit reverses the changes in
https://github.com/OpenVAA/voting-advice-application/commit/4ee7ada8157c8bf918815314f4faa6b4f440effe