feat: add `cancel` action to `Banner`
Add a `cancel` action to `Banner` which can be used to execute an
arbitrary callback, e.g., to cancel the pre-registration process.

The button is very similar to `return` but is a distinct one, because
the former offers some (now updated) default actions and its behaviour
may change in the future.

To test, add the following on any route:

```ts
  import { getLayoutContext } from '$lib/contexts/layout';
  import { onDestroy } from 'svelte';
  const { topBarSettings } = getLayoutContext(onDestroy);
  topBarSettings.push({
    actions: {
      cancel: 'show',
      cancelButtonLabel: $t('common.cancel'),
      cancelButtonCallback: function() {
        console.error('CANCEL!')
      },
    }
  });
```