feat: option for selecting `Constituency` before `Elections`
Add a `elections.startFromConstituencyGroup` setting that can be used to
select a `ConstituencyGroup` from which a `Constituency` must be picked
before moving to `Election` selection.

Supporting this requires changes to:
- utils for implying `constituencyId`s
- rerouting checks on the `/elections`, `/constituencies` and
  `/(located)` Voter App routes
- the order of items shown on the `/intro` route
- the order of items in `VoterNav`

The setting is useful in apps which contain multiple `Election`s which
are applicable to only some `Constituency`s in a common hierarcy. An
example are the Finnish local elections, which have two sub-elections:
regional ones which are only held on 21 regions outside the capital and
municipal ones which are held in all the country’s municipalities.

The regions can be implied from the municipalities, but in the case of
the capital Helsinki, no such region exists. This means that it's better
to let the Voter first choose their home municipality and only the
election selector if they are in a region that will host both elections.