feat: `Question.customData.hidden` property
If `true`, the `Question` is hidden in the Voter App but still visible
in the Candidate App.

The check for this has to be unsatisfactorily be done in both the
`questionStore`s in the contexts and in `EntityDetails`, because we want
to hide it in both the details view and in filters.