fix: `matchStore` error when there are no nominations
Return an empty array when there are nominations instead of calling the
`match` function which throws when there are no targets.

Nominations may be zero in cases where some constituencies have no
participating candidates or parties. The results page will anyway show a
notice when there are no entities, so throwing an error is not
necessary.