fix: unsupported adapter error when using `local` adapter
Only throw an error when attempting to use an unsupported `DataWriter`,
not immediately upon module import.

This is needed to prevent errors when using the `local` data adapter,
even when only accessing the Voter App, because it conditionally uses the
`CandidateContext`. Due to the import logic, the context module is
imported and `$lib/api/dataWriter` along with it, even though the
context is only instantiated when accessing the Candidate App, which the
`local` adapter does not support.