feat: Admin App
Add a scaffold for the Admin App on the `/admin` route.

Add `AdminContext` for use with the app.

Reorganize common translations used in both Candidate and Admin Apps.

Within the Admin App, add an incomplete draft of the argument
condensation function.

Add preliminary translations for some parts of the Admin App, which are
not yet available.