fix: available nominations check
If some nominations seem not to be available when loading them on the
`(located)` route, recheck after a small timeout to allow for the
`$nominationsAvailable` store to be updated which may not happen even
though we wait for one `tick()`.

This solution is imperfect and could be changed to one that explicitly
waits for an updated event or similar from the store or `DataRoot`.