feat[backend][frontend]: `Candidate.termsOfUseAccepted` property
Add a property for recording the datetime of a candidate’s acceptance of
the terms of use, which is `null` if not yet accepted.

This is automatically set by the `/auth/candidate/preregister` route,
because the preregistration process includes accepting the terms.

The `termsOfUseAccepted` property is returned by the `DataWriter`
methods dealing with candidate date, but not the voter-facing ones in
`DataProvider`. The latter, however, filters out candidates who have not
accepted the terms.

NB. A more consistent place for this property might be the `User`
object, but this would involve an extra populate call when fetching the
candidates.