refactor[app-shared][frontend]: localized data types
Refactor and centralize all definitions for the localized, i.e.
multi-locale, extended data types in the `@openvaa/app-shared` module.

These are the same types that have already been used in Strapi to store
such data and Strapi-specific definitions are thus removed.

The utilities for translating localized types are refactored and moved
from Strapi-specific utils to general ones.

NB. In the post-Strapi future, the whole app-shared module may become
obsolete and the types will then be moved to the frontend package.