feat: `hero` `CustomData` property and `Hero` component
Add `hero` property to `Question`s’  and `QuestionCategory`s’
`CustomData` type, replacing the latters `emoji` property.

The property is of type `HeroContent` which may contain either an
`Emoji` or an `Image` (from `@openvaa/data`).

Add a `Hero` component for displaying the content.

Apply the component on the question and question category routes.

Add and extend existing api utilities for parsing the `hero` property.