feat[backend]: Admin Job content type
Add Admin Job content type to Strapi for saving job results.

Add `insertJobResult` method to `DataWriter` and implement it in
`StrapiDataWriter`.

Add a utility method with the same name to `AdminContext`.

NB. The backend supports reading and updating Admin Jobs but in the
frontend only the insert method is now implemented.