feat: disable filtering of parent nominations with no children
The filtering does not work with the current setup and should be handled
already on the server or by the data provider.

As this is not a crucial issue, it will be postponed until the backend
is updated. In the meantime, purged parties with no candidates will just
appear as empty parties in the party results.